#8.14 2 bugs fix. 1. password 저장하는 것이 아니라면 해싱 방지, 2. 저자가 다른 사용자가 비디오를 수정할 수 없게 방지

+++ src/controllers/videoController.js
@@ -28,16 +28,28 @@ export const watch = async (req, res) => {
 	return res.render("Watch", { pageTitle: video.title, video });
 };
 export const getEdit = async (req, res) => {
+	const {
+		user: { _id },
+	} = req.session;
+	//req.session.user._id
 	const { id } = req.params;
 	const video = await Video.findById(id);
-	console.log("getEdit: ", video);
+	// console.log("getEdit: ", video);
 	if (!video) {
 		return res.status(404).render("404", { pageTitle: "Video not found" });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
+		//403은 forbidden을 의미한다.
+		//비디오를 업로드한 사용자 본인이 아니라면 접속할 수 없게 설정.
 	}
 	return res.render("Edit", { pageTitle: `Editing`, video });
 };
 
 export const postEdit = async (req, res) => {
+	const {
+		user: { _id },
+	} = req.session;
 	const { id } = req.params;
 	const { title, description, hashtags } = req.body;
 
@@ -45,6 +57,9 @@ export const postEdit = async (req, res) => {
 	console.log("postEdit: ", video);
 	if (!video) {
 		return res.render("404", { pageTitle: "Video not found" });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
 	}
 	await Video.findByIdAndUpdate(id, {
 		title,
@@ -96,15 +111,28 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+	const {
+		user: { _id },
+	} = req.session;
 	const { id } = req.params;
-	//delete video.
-	//여기서 작성자만 삭제할 수 있게 변경이 필요한데, 추후 강의 진행하며 진행 필요.
+
+	const video = await Video.findById(id);
+	//비디오가 없다면 404로 넘어가고, 있다면 로그인한 사람과 비디오의 owner가 같은지 체크해서, 다르다면 forbidden하여 홈으로 돌려보낸다.
+	if (!video) {
+		return res.render("404", { pageTitle: "Video not found" });
+	}
+	if (String(video.owner) !== String(_id)) {
+		return res.status(403).redirect("/");
+	}
+
 	try {
+		//video를 삭제하는 부분
 		const result = await Video.findByIdAndDelete(id);
 		console.log("this is result", result);
 	} catch (error) {
 		console.log(error);
 	}
+
 	return res.redirect("/");
 };
